# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Blink HTTP API
  description: |
    本仓库 HTTP 接口的 OpenAPI 描述；与 `infrastructure/interface/http/` 及 `cmd/main.go` 中的路由一致。

    **`POST /auth/register`**：始终挂载（见 `cmd/main.go`），用于邮箱密码注册并建立 `builtin` 身份绑定。

    **内置 IdP**：`/auth/idp/*` 以及 OAuth 提供方 **`builtin`** 仅在配置了 `BLINK_PUBLIC_BASE_URL`、`BLINK_OAUTH_CLIENT_SECRET` 等环境变量时启用；未配置时上述路径返回 **404**，但不影响 `/auth/register`。

    **代码生成**：仅带标签 `servergen` 的操作会由 oapi-codegen 生成 Gin 注册代码与 `ServerInterface` 方法；其余路径为手写路由，此处仅作文档。

    **社交与管理**：`/api/*`、`/admin/api/*` 为手写路由（见 `cmd/main.go`），OpenAPI 中 `social` / `admin` / `feedback` / `notifications` 标签描述之。

    **管理权限**：`users.role` 为 `admin` 或 `super_admin` 时可访问 `/admin/api/*`（需有效会话）。部分能力（如 SMTP 配置）在实现上仍建议仅授予 `super_admin`。
  version: 0.1.0
servers:
  - url: http://127.0.0.1:11110
    description: 本地开发（默认 BLINK_HTTP_ADDR=:11110）
  - url: http://localhost:11110
    description: 本地开发（localhost）
  - url: "http://127.0.0.1:{port}"
    description: 本地其它端口（与当前 BLINK_HTTP_ADDR 一致，在 Swagger 里改 port）
    variables:
      port:
        default: "11110"
  - url: /
    description: 与 Swagger UI 同源（把文档托管在本 API 上时用此项）
...
paths:
...
  /api/posts:
    get:
      operationId: listPosts
      tags: [social]
      summary: 公开帖子流
      description: |
        Query `category_id`、`uncategorized=1`、`cursor`（上一页最后一条 id）、`limit`。
        可选会话（Cookie / Bearer）：登录后响应中每条帖子含 `liked`；`like_count` 始终返回。
      responses:
        "200":
          description: '`{ "posts": [...], "next_cursor": int64|null }`'
    post:
      operationId: createPost
      tags: [social]
      summary: 发帖（需登录）
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                category_id:
                  type: integer
                  format: int64
                  nullable: true
                images:
                  type: array
                  items:
                    type: string
                draft:
                  type: boolean
      responses:
        "201":
          description: 帖子 JSON
  /api/posts/{id}:
    get:
      operationId: getPost
      tags: [social]
      summary: 帖子详情（可选登录；作者可看草稿/非公开）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 帖子 JSON
        "404":
          description: 不存在或对当前访客不可见
    patch:
      operationId: patchPost
      tags: [social]
      summary: 修改自己的帖子
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                category_id:
                  type: integer
                  format: int64
                clear_category:
                  type: boolean
                images:
                  type: array
                  items:
                    type: string
                status:
                  type: integer
      responses:
        "200":
          description: 更新后帖子
    delete:
      operationId: deletePost
      tags: [social]
      summary: 删除自己的帖子（软删）
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
  /api/me/posts:
    get:
      operationId: listMyPosts
      tags: [social]
      summary: 我的帖子
      security:
        - sessionCookie: []
        - bearerAuth: []
      description: Query `include_draft=1`、`cursor`、`limit`。
      responses:
        "200":
          description: 同 listPosts
...
  /api/posts/{id}/replies:
    get:
      operationId: listReplies
      tags: [social]
      summary: 评论列表（仅已发布且公开的根帖）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: '`{ "replies": [...], "next_cursor" }`'
    post:
      operationId: createReply
      tags: [social]
      summary: 发表评论
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                parent_reply_id:
                  type: integer
                  format: int64
                  nullable: true
      responses:
        "201":
          description: 评论 JSON
...
  /api/replies/{id}/likes:
    get:
      operationId: getReplyLikes
      tags: [social]
      summary: 评论点赞数（可选登录返回 liked）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: '`{ "reply_id", "like_count", "liked"? }`'
        "404":
          description: 评论不存在
  /api/replies/{id}/like:
    post:
      operationId: likeReply
      tags: [social]
      summary: 点赞评论
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
        "409":
          description: 已点赞
    delete:
      operationId: unlikeReply
      tags: [social]
      summary: 取消评论点赞
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
        "404":
          description: 未点赞
  /api/replies/{id}:
    delete:
      operationId: deleteReply
      tags: [social]
      summary: 删除自己的评论
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    sessionCookie:
      type: apiKey
      in: cookie
      name: blink_session
...